requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
basedosdados==2.0.2
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.5
db-dtypes==1.5.0
distro==1.9.0
docopt==0.6.2
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.2
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.40.1
google-cloud-bigquery-connection==1.20.0
google-cloud-bigquery-storage==2.36.1
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-genai==1.73.1
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
granian==2.7.1
greenlet==3.3.1
grip==4.6.2
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loguru==0.7.3
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.16.0
numpy==2.4.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandas-gbq==0.33.0
path-and-address==2.0.1
pillow==12.1.1
platformdirs==4.9.1
plotly==6.5.2
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.3
pytest-cov==7.1.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2025.2
redis==7.1.1
referencing==0.37.0
reflex==0.8.26
reflex-hosting-cli==0.1.61
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.52.1
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tomlkit==0.11.8
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
win32_setctime==1.2.0
wrapt==2.1.1
wsproto==1.3.2
